Move chain_analysis, syntaxerror, trace_cpy into trace package; keep inspector at top level.
- Relocate traceback-specific helpers into tracerite/trace/.
- Keep tracerite.inspector at the top level; trace.finalize imports it
  absolutely.
- Use absolute imports (tracerite.logging) for cross-package dependencies
  inside trace/.
- Update pyproject.toml coverage omit for the moved trace_cpy path.
- Fix tests and monkeypatch paths for the new module locations.
