+++ tests/test_html_cornercases.py
@@ -5,9 +5,9 @@ import sys
 import pytest
 from bs4 import BeautifulSoup
 
-from tracerite.chain_analysis import build_chronological_frames
 from tracerite.html import html_traceback, javascript, style
 from tracerite.trace import extract_chain, extract_exception, symbols
+from tracerite.trace.chain_analysis import build_chronological_frames
 
 
 def _chrono(exc_info):
